pyproject.toml
[tool.poetry]
name = "apriltag-pygui"
version = "0.1.0"
description = "A Simple GUI to edit apriltags"
authors = ["Estevao Santos <estevaojs@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Pillow = "^9.1.1"
PySimpleGUI = "^4.60.1"
loguru = "^0.6.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
